even more css, html, and images
I was able to create a button that links to the admin.html page without it looking weird on the login page. I created a new css file called visit-browse.css that connects to the visitorBrowse.html page. I also created graphs based on Rick and Morty that focus on views by season, by age group, and by gender so far. Lastly i created a view-graph.html page on the admin side.